Seperate ResourceOperationListItem
